A few tests for the Add Job forms. Correcting Integrity Error
When adding tests to the Add Job forms, I've noticed a bug: when a user submits the form
with an existing company but with a different website, we've got the Integrity Error.

I've corrected it by adding a bit of logic: when that happens, the user gets a validation
error and they can choose between keeping the old website or overwriting it with the new
one.

I've added tests for the following paths:
- submitting the form with a new company
- submitting the form for an existing company with the same website
- submitting the form for an existing company but with a different website (validation)
- user chooses to overwrite the old website
- user chooses to keep the old website

Some more tests are needed but didn't want to make this pull request too big.